+++ CHANGELOG.md
@@ -38,6 +38,8 @@ Code style revamp! We reworked how routines, routine methods, properties, and st
 
 Added a new preference for notebooks called "Quiet Mode" that allows you to control the IDL preference for `!quiet` when executing notebook cells.
 
+When using ".edit" in the debug console, added a message that let's you know that a matching file was not found.
+
 ## 4.1.0 November 2023
 
 Fixed a bug with the debugger that would fail to get information about your IDL session.